+++ popup/templates/extension-rate.js
@@ -1,0 +1,13 @@
+document.write(/*html*/`
+<div id="extension-rate">
+	<div class="rate-subtext">Please, consider giving me a review!</div>
+	<div class="rate-stars">
+		<a href="https://chromewebstore.google.com/detail/wanikani-kanji-highlighte/pdbjikelneighjgjojikkmhiehpcokjm/reviews#:~:text=Write%20a%20Review" target="_blank" class="clickable icon" title="Rate 1 star"><img src="/images/star.png" alt="star"></a>
+		<a href="https://chromewebstore.google.com/detail/wanikani-kanji-highlighte/pdbjikelneighjgjojikkmhiehpcokjm/reviews#:~:text=Write%20a%20Review" target="_blank" class="clickable icon" title="Rate 2 stars"><img src="/images/star.png" alt="star"></a>
+		<a href="https://chromewebstore.google.com/detail/wanikani-kanji-highlighte/pdbjikelneighjgjojikkmhiehpcokjm/reviews#:~:text=Write%20a%20Review" target="_blank" class="clickable icon" title="Rate 3 stars"><img src="/images/star.png" alt="star"></a>
+		<a href="https://chromewebstore.google.com/detail/wanikani-kanji-highlighte/pdbjikelneighjgjojikkmhiehpcokjm/reviews#:~:text=Write%20a%20Review" target="_blank" class="clickable icon" title="Rate 4 stars"><img src="/images/star.png" alt="star"></a>
+		<a href="https://chromewebstore.google.com/detail/wanikani-kanji-highlighte/pdbjikelneighjgjojikkmhiehpcokjm/reviews#:~:text=Write%20a%20Review" target="_blank" class="clickable icon" title="Rate 5 stars"><img src="/images/star.png" alt="star"></a>
+	</div>
+	<div class="rate-subtext">Having issues using the extension?<br>Let me know in the <a href="https://community.wanikani.com/t/browser-extension-wanikani-kanji-highlighter-2021/50455" target="_blank">Forum</a>, via <a href="mailto:diogocorreia.dev@gmail.com?subject=Issue with Wanikani Kanji Highlighter">Email</a> or on <a href="https://github.com/digas99/wanikani-kanji-highlighter/issues" target="_blank">Github</a>.</div>
+</div>
+`);